+++ xcode/config.h
@@ -8,7 +8,7 @@
 //
 
 // Version numbers
-#define PAPPL_VERSION		"1.0b2"
+#define PAPPL_VERSION		"1.0rc1"
 #define PAPPL_VERSION_MAJOR	1
 #define PAPPL_VERSION_MINOR	0