+++ CHANGES.md
@@ -31,6 +31,8 @@ Changes in v1.0b2
 - `papplDeviceRead` now has a 100ms timeout for USB and network connections.
 - Implemented back-channel and status updates for the USB printer gadget.
 - Fixed a bug in the log rotation code.
+- Fixed some threading bugs with the various object lists managed by the
+  system.
 
 
 Changes in v1.0b1