+++ CHANGES.md
@@ -30,6 +30,8 @@ Changes in v1.0b2
   registration for the system web page in multi-queue mode.
 - `papplDeviceRead` now has a 100ms timeout for USB and network connections.
 - Implemented back-channel and status updates for the USB printer gadget.
+- Finished implementation of test suite for major code paths/job processing
+  functionality.
 - Fixed a bug in the log rotation code.
 - Fixed some threading bugs with the various object lists managed by the
   system.