+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in v1.0b2
 - Now check whether a named printer already exists (Issue #76)
 - Support for "output-bin-default" was missing from the web interface
   (Issue #77)
+- Fixed support for vendor options at the command-line (Issue #79)
 - The main loop now shows an error message if an option is provided after "-o"
   without a space (Issue #80)
 - Fixed test page and identify buttons (Issue #81)