+++ CHANGES.md
@@ -2,8 +2,8 @@ Changes in PAPPL
 ================
 
 
-Changes in v1.0b2
------------------
+Changes in v1.0rc1
+------------------
 
 - Added IEEE-1284 device ID to argument list for printer driver callbacks
   (Issue #70)